Rename color variant variables
Using letters instead of numbers to differentiate between variant variables allows for definitions that are easier to understand if tokens are to be employed. A declaration such as `$line-height-small-1: $line-height-3` can be somewhat confusing because different numbers are used to express the same characteristic. `$line-height-small-a: $line-height-3` creates a distinction, implying that each variable has a different purpose.

This naming convention should be employed for all variables.