Revise comments on variant variables
Naming color variations based on their relation to regular colors could be problematic because the latter may have very low or high lightness. This will lead to variations that are designated as light or dark even though they aren't. Instead, colors should be labeled as dark, medium, and light based on their lightness falling somewhere between 0 and 39, 40 and 59, or 60 and 100, respectively.

Unlike colors, which have a maximum lightness of 100, variables that establish border width, space, font size, and line height variations take on values from a set lacking an upper boundary so they should be defined as large or small with respect to a regular value.