Revise button hover and focus states
Signaling focus with the aid of outlines or shadows may lead to item overlap. This can be displeasing when navigating a group of buttons or a text input and a button placed next to each other. Background color changes are an adequate way of highlighting focus if it happens via mouse or touch. Outlines should be employed to show when an element receives keyboard focus once `:focus-visible` is supported by major browsers.