Remove disabled state class
Classes only apply styles; they do not remove functionality. Items with `.is-disabled` would look disabled but behave normally. The `disabled` attribute is well-suited for this job on form elements but is not valid on links. However, links are meant to act as a means of perpetual access to some other resource so should never be disabled.