Use pointer cursor
After further consideration it does seem that there needs to be some visual feedback when hovering over interactive form elements. Setting default cursor styles for all appropriate elements via the reset is not possible because there is no way of selecting the exact labels associated with a disabled item so styles must be applied at the component level.